client_requirements.txt
Pillow
gradio==4.16.0
pydantic==2.6.0
Requests==2.31.0